Calendar: support delete only one instance of reoccuring events.

+++ app/actions/CalendarActions.js
@@ -40,17 +40,27 @@ export default {
         }
     },
 
-    deleteNoRepeatEvent(id, success) {
+    deleteEvent(data, success) {
+        let id = data.id;
+        let repeatedNumber = data.repeatedNumber;
+
+        if (id === null || id === undefined) {
+            return;
+        }
+
+        let url = `/platform/api/events/${id}`;
+        if (repeatedNumber) {
+            url += `/${repeatedNumber}`;
+        }
+
         $.ajax({
-            url: "/platform/api/events/" + id,
+            url: url,
             type: "delete"
         }).done(() => {
             //console.log("haha");
             AppDispatcher.dispatch({
                 type: CalendarActionTypes.DELETE_EVENT,
-                data: {
-                    id: id
-                }
+                data: data
             });
             if (success && typeof success === "function") {
                 success();
